Make the Gradle build idiomatic
This commit refactors the build to make it Gradle idiomatic.
This greatly simplifies the setup, and removes a bunch of errors
in configuration.

There is now a task called `publishAllPublicationsToBuildRepository`
which creates a Maven repository under `build/repo` to check what
is going to be published.

The version of the project is managed under `gradle/libs.versions.toml`