IDs should be String (#54)
* IDs should be String

According to Mastodon API, various ID values are "cast from an integer, but not guaranteed to be a number". This changes entities and related calls from Long to String id values.

* IDs should be String

According to Mastodon API, various ID values are "cast from an integer, but not guaranteed to be a number". This changes entities and related calls from Long to String id values.

* IDs should be String

According to Mastodon API, various ID values are "cast from an integer, but not guaranteed to be a number". This changes entities and related calls from Long to String id values.

Co-authored-by: André Gasser <andre.gasser@protonmail.com>