Enable local, remote or combined results when getting timelines
Public.getPublic and Public.getTag now have a statusOrigin parameter that can be set to different values defined in Public.StatusOrigin to get either local or remote results, or a combination of both.

Removing previously used separate functions that did some of the above.

Documenting all of Public.kt with proper KDoc.

Closes #75.