don't expose okhttp3.MultipartBody.Part via Media.postMedia()
Changes Media.postMedia() to accept a File and a media type String as parameters, then builds the necessary okhttp3.MultipartBody.Part itself instead of exposing this.

A more extensive Builder was suggested in #62, but I think that would need to live elsewhere, not directly in one of our method classes.