remove redirect URI default value from methods
Default value was the out-of-band URI "urn:ietf:wg:oauth:2.0:oob", which is not useful in most production contexts. Keeping it as a default might lead to hard-to-debug situations where an actually sensible parameter is omitted and authentication fails as a result of this. The out-of-band URI can still be used, but needs to be specified.